requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
playwright>=1.35.0,<1.40.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
nuitka>=1.8.0

